% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUpset.R
\name{addColKey}
\alias{addColKey}
\title{Defines function for plotting color key.}
\usage{
addColKey(name, colfn)
}
\arguments{
\item{name}{Color key title.}

\item{colfn}{Function for mapping colors.}
}
\value{
An \link{\code{AnnotationFunction}}.
}
\description{
Defines function for plotting color key.
}
