% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmfns.R
\name{cond.col}
\alias{cond.col}
\title{Color scale for categorical data.}
\usage{
cond.col(cond, colfn = rainbow, ...)
}
\arguments{
\item{cond}{A vector that can be coerced to a factor.}

\item{colfn}{A function that returns a color map for each level in \code{cond}.}

\item{...}{Additional arguments to \code{colfn}.}
}
\value{
A named vector of colors corresponding to the levels of \code{cond}.
}
\description{
Color scale for categorical data.
}
