% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmdot.R
\name{dotPscale}
\alias{dotPscale}
\title{Version of hyperDot using p-value instead of number of replicates for dot size.
accepts the results of an enrichment test applied to each cell in a matrix
and writes a dotplot of the results}
\usage{
dotPscale(
  odds,
  fdr,
  q,
  plim = 0.01,
  outl.name = "size",
  size.name = "-log10(FDR)",
  ...
)
}
\arguments{
\item{odds}{A matrix of log2 odds ratios from an enrichment test.}

\item{fdr}{A matrix of FDR values from an enrichment test.}

\item{q}{A matrix of test counts for each test.}

\item{plim}{Maximum p-value on the size scale. Values above \code{plim} are set to \code{plim}.}

\item{...}{Additional arguments to \code{hmdot()}.}
}
\description{
Version of hyperDot using p-value instead of number of replicates for dot size.
accepts the results of an enrichment test applied to each cell in a matrix
and writes a dotplot of the results
}
