% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUpset.R
\name{getOR}
\alias{getOR}
\title{Runs binomial of a set intersection vs. independence}
\usage{
getOR(x, id, mat, N = NA)
}
\arguments{
\item{x}{The intersect size.}

\item{id}{The intersect names.}

\item{mat}{The logical matrix of intersections.}

\item{N}{The universe size.}
}
\value{
A numeric vector containing the log2 odds ratio and p-value of binomial test.
}
\description{
Runs binomial of a set intersection vs. independence
}
