% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmfns.R
\name{hm.cell}
\alias{hm.cell}
\title{Wrapper for \code{Heatmap()} which allows specifying cell dimensions and resizing the heatmap accordingly.}
\usage{
hm.cell(
  x,
  ...,
  cell.h = NULL,
  cell.w = NULL,
  height = NULL,
  width = NULL,
  units = "in"
)
}
\arguments{
\item{x}{A numeric matrix to be plotted.}

\item{...}{Additional arguments to \code{Heatmap()}.}

\item{cell.h}{The cell height.}

\item{cell.w}{The cell width.}

\item{height}{The heatmap height. Ignored if \code{cell.h} is specified.}

\item{width}{The heatmap width. Ignored if \code{cell.w} is specified.}

\item{units}{The unit scale to be used for \code{cell.h} and \code{cell.w}.}
}
\value{
A ComplexHeatmap.
}
\description{
Wrapper for \code{Heatmap()} which allows specifying cell dimensions and resizing the heatmap accordingly.
}
