% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmfns.R
\name{hm.char}
\alias{hm.char}
\title{Wrapper for \code{Heatmap()} which allows specifying character width for row and column names and resizing the heatmap accordingly.}
\usage{
hm.char(
  x,
  ...,
  char.h = NULL,
  char.w = NULL,
  legend.w = 0,
  height = NULL,
  width = NULL,
  heatmap_height = NULL,
  heatmap_width = NULL,
  units = "in"
)
}
\arguments{
\item{x}{A numeric matrix to be plotted.}

\item{...}{Additional arguments to \code{Heatmap()}.}

\item{char.h}{The characther width for column names.}

\item{char.w}{The character width for row names.}

\item{legend.w}{Buffer width for legend.}

\item{heatmap_height}{The heatmap height. Ignored if \code{cell.h} is specified.}

\item{heatmap_width}{The heatmap width. Ignored if \code{cell.w} is specified.}

\item{units}{The unit scale to be used for \code{cell.h} and \code{cell.w}.}
}
\value{
A ComplexHeatmap.
}
\description{
Wrapper for \code{Heatmap()} which allows specifying character width for row and column names and resizing the heatmap accordingly.
}
