% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmdot.R
\name{hmdot}
\alias{hmdot}
\title{accepts the results of an enrichment test applied to each cell in a matrix
and writes a dotplot of the results}
\usage{
hmdot(
  odds,
  fdr,
  q,
  file,
  path = ".",
  cell.dim = 0.15,
  width = 12,
  height = 12,
  ...
)
}
\arguments{
\item{odds}{A matrix of log2 odds ratios from an enrichment test.}

\item{fdr}{A matrix of FDR values from an enrichment test.}

\item{q}{A matrix of test counts for each test.}

\item{file}{The output file name.}

\item{path}{The output file path.}

\item{cell.dim}{The width & height of each heatmap cell in inches.}

\item{width}{The width of the output device in inches.}

\item{height}{The height of the output device in inches.}

\item{...}{Additional arguments to \code{hm.cell()}.}
}
\description{
accepts the results of an enrichment test applied to each cell in a matrix
and writes a dotplot of the results
}
