% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyper.R
\name{phyper2}
\alias{phyper2}
\title{Two-tailed version of \code{\link{phyper}}. It simply takes the minimum of the upper and lower tails and multiplies the result by 2.}
\usage{
phyper2(...)
}
\arguments{
\item{...}{Arguments to \code{phyper()}.}
}
\value{
A p-value.
}
\description{
Two-tailed version of \code{\link{phyper}}. It simply takes the minimum of the upper and lower tails and multiplies the result by 2.
}
\seealso{
\code{\link{phyper}}
}
