% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmcor.R
\name{quantHeatmap}
\alias{quantHeatmap}
\title{Wrapper function for \code{\link{writeHeatmap}}.}
\usage{
quantHeatmap(
  mat,
  filename,
  colfn = col.z,
  quant = 0.01,
  cell.w = 0.2,
  cell.h = 0.05,
  char.dim = 0.1,
  buffer.h = 5,
  buffer.w = 5,
  ...
)
}
\arguments{
\item{mat}{A numeric matrix.}

\item{filename}{The output file name.}

\item{colfn}{Function to generate color scale. By default uses \link{\code{col.abs()}}.}

\item{quant}{Quantile used to set color scale limits.}

\item{cell.w}{Width of heatmap cells.}

\item{cell.h}{Height of heatmap cells.}

\item{char.dim}{Character width for row and column names.}

\item{...}{Additional arguments to \link{\code{Heatmap()}}.}
}
\value{
A Heatmap object.
}
\description{
Wrapper function for \code{\link{writeHeatmap}}.
}
