% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUpset.R
\name{upsetAnnoBar}
\alias{upsetAnnoBar}
\title{Plot bars of UpSet.}
\usage{
upsetAnnoBar(..., gp = gpar(fill = "black"))
}
\arguments{
\item{...}{Arguments to \link{\code{ComplexHeatmap::anno_barplot()}}.}

\item{gp}{A \link{\code{grid::gpar()}} object.}
}
\value{
A \link{\code{HeatmapAnnotation}} object.
}
\description{
Plot bars of UpSet.
}
