% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmfns.R
\name{col.z}
\alias{col.z}
\title{Color scale for a specified quantile. This scale is intended for heatmaps containing negative and positive values, so the range is set to \code{c(quant,1-quant)}.}
\usage{
col.z(x, quant = 0.01, mid = 0, cols = c("blue", "white", "red"))
}
\arguments{
\item{x}{A numeric matrix or vector.}

\item{quant}{The quantile to be used as the scale limits.}

\item{mid}{The midpoint of the scale.}
}
\value{
A \code{colorRamp2} scale which can be passed to \code{Heatmap()}.
}
\description{
Color scale for a specified quantile. This scale is intended for heatmaps containing negative and positive values, so the range is set to \code{c(quant,1-quant)}.
}
