% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmdot.R
\name{hmdot}
\alias{hmdot}
\title{accepts the results of an enrichment test applied to each cell in a matrix
and writes a dotplot of the results}
\usage{
hmdot(
  mat,
  outl,
  size,
  col.mat,
  col.outl,
  scale,
  mat.name = "log2(OR)",
  outl.name = "-log10(FDR)",
  size.name = "size",
  file,
  path = ".",
  cell.dim = 0.15,
  width = 12,
  height = 12,
  append.date = F,
  ...
)
}
\arguments{
\item{mat}{A matrix of values shown by the dot color.}

\item{outl}{A matrix of values shown by the dot outlline.}

\item{size}{A matrix of values shown by the dot size.}

\item{col.mat}{A color scale for \code{mat}.}

\item{col.outl}{A color scale for \code{outlline}.}

\item{scale}{A vector of length 2 giving the min and max values to scale the size of the dots.}

\item{file}{The outlput file name.}

\item{path}{The outlput file path.}

\item{cell.dim}{The width & height of each heatmap cell in inches.}

\item{width}{The width of the outlput device in inches.}

\item{height}{The height of the outlput device in inches.}

\item{...}{Additional arguments to \code{hm.cell()}.}
}
\description{
accepts the results of an enrichment test applied to each cell in a matrix
and writes a dotplot of the results
}
